+++ src/main/java/com/verf_system/verfS/database/repository/IItensReceitaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IItensReceitaRepository extends JpaRepository<ItensReceitaEntity, Integer> {
+public interface IItensReceitaRepository extends JpaRepository<ItensReceitaEntity, Long> {
 }
