+++ src/main/java/com/verf_system/verfS/database/repository/ITintaRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ITintaRepository extends JpaRepository<TintaEntity, Integer> {
+public interface ITintaRepository extends JpaRepository<TintaEntity, Long> {
 
 
 }
